api/requirements.txt
Django==5.0.4
msal==1.29.0
PyJWT==2.8.0
python-dotenv==1.0.1
django-cors-headers==4.4.0